web/requirements.dev.txt
django==4.0
django-otp==1.1.3
djangorestframework==3.13.1
gunicorn==20.1.0
requests==2.27.1
django-dotenv==1.4.2
psycopg2-binary==2.9.3
django-storages==1.12.3
boto3==1.24.30
qrcode==7.3.1
flake8==4.0.1